move device-conect-state to types folder
