+++ src/electron/flux/types/device-connect-state.tsx
