+++ src/electron/views/device-connect-view/components/device-connect-body.tsx
@@ -4,12 +4,12 @@ import { NamedFC } from 'common/react/named-fc';
 import { BrowserWindow } from 'electron';
 import * as React from 'react';
 
+import { DeviceConnectState } from '../../../flux/types/device-connect-state';
 import { deviceConnectBody } from './device-connect-body.scss';
 import { DeviceConnectConnectedDevice } from './device-connect-connected-device';
 import { DeviceConnectFooter, DeviceConnectFooterDeps } from './device-connect-footer';
 import { DeviceConnectHeader } from './device-connect-header';
 import { DeviceConnectPortEntry, DeviceConnectPortEntryDeps, DeviceConnectPortEntryViewState } from './device-connect-port-entry';
-import { DeviceConnectState } from './device-connect-state';
 
 export type UpdateStateCallback = (newState: DeviceConnectState, deviceName?: string) => void;
