+++ src/electron/views/device-connect-view/components/device-connect-port-entry.tsx
@@ -6,8 +6,8 @@ import { MaskedTextField } from 'office-ui-fabric-react/lib/TextField';
 import * as React from 'react';
 
 import { DeviceConnectActionCreator } from '../../../flux/action-creator/device-connect-action-creator';
+import { DeviceConnectState } from '../../../flux/types/device-connect-state';
 import { deviceConnectPortEntry, portNumberField } from './device-connect-port-entry.scss';
-import { DeviceConnectState } from './device-connect-state';
 
 export type DeviceConnectPortEntryViewState = {
     deviceConnectState: DeviceConnectState;
