+++ src/tests/unit/tests/electron/views/device-connect-view/components/device-connect-connected-device.test.tsx
@@ -4,7 +4,7 @@ import {
     DeviceConnectConnectedDevice,
     DeviceConnectConnectedDeviceProps,
 } from 'electron/views/device-connect-view/components/device-connect-connected-device';
-import { DeviceConnectState } from 'electron/views/device-connect-view/components/device-connect-state';
+import { DeviceConnectState } from 'electron/flux/types/device-connect-state';
 import { shallow } from 'enzyme';
 import * as React from 'react';
