+++ src/tests/unit/tests/electron/views/root-container/components/root-container.test.tsx
@@ -3,7 +3,7 @@
 import { BaseClientStoresHub } from 'common/stores/base-client-stores-hub';
 import { ClientStoresHub } from 'common/stores/client-stores-hub';
 import { ScanStatus } from 'electron/flux/types/scan-status';
-import { DeviceConnectState } from 'electron/views/device-connect-view/components/device-connect-state';
+import { DeviceConnectState } from 'electron/flux/types/device-connect-state';
 import {
     RootContainer,
     RootContainerDeps,
